docs: MemberConverter 통합
