+++ Leon/umc10th/src/main/java/com/example/umc10th/domain/member/controller/UserController.java
@@ -29,14 +29,19 @@ public class UserController {
         return ApiResponse.onSuccess(code, userService.getMyInfo(memberId));
     }
 
-    @GetMapping("/me/missions")
+    @PostMapping("/me/missions")
     public ApiResponse<Page<MemberResponseDTO.MyMission>> getMyMissions(
+            @RequestBody MemberRequestDTO.MyMissionRequest request,
             @RequestParam(required = false) MissionStatus status,
             @PageableDefault(size = 10) Pageable pageable
     ) {
-        Long memberId = 1L;
+
         BaseSuccessCode code = MemberSuccessCode.OK;
-        return ApiResponse.onSuccess(code, userService.getMyMissions(memberId, status, pageable));
+        return ApiResponse.onSuccess(code, userService.getMyMissions(
+                request.getMemberId(),
+                status,
+                pageable
+        ));
     }
 
     @GetMapping("/me/reviews")
