+++ Leon/umc10th/src/main/java/com/example/umc10th/domain/member/service/impl/AuthServiceImpl.java
@@ -27,7 +27,7 @@ public class AuthServiceImpl implements AuthService {
 
         Member savedMember = memberRepository.save(member);
 
-        return memberConverter.toGetMyInfo(savedMember);
+        return memberConverter.toMyInfo(savedMember);
     }
 
     @Override
